feat: add select asset page
