+++ components/send/asset-selector/Erc20SelectItem.tsx
@@ -4,18 +4,11 @@ import { useNavigate } from "react-router-dom";
 import { getChainImage } from "@/lib/layer2";
 import useWalletManager from "@/hooks/useWalletManager";
 import { twMerge } from "tailwind-merge";
-import { useFormContext } from "react-hook-form";
 import useErc20Balance from "@/hooks/evm/useErc20Balance";
 import { Erc20Asset } from "@/contexts/EvmAssets";
-import { formatToken } from "@/lib/utils.ts";
+import { formatToken, textEllipsis } from "@/lib/utils.ts";
 
-export default function Erc20SelectItem({
-  asset,
-  toggleShow,
-}: {
-  asset: Erc20Asset;
-  toggleShow: () => void;
-}) {
+export default function Erc20SelectItem({ asset }: { asset: Erc20Asset }) {
   const { wallet } = useWalletManager();
   const navigate = useNavigate();
   const { data } = useErc20Balance(
@@ -24,24 +17,11 @@ export default function Erc20SelectItem({
     asset.chainId,
   );
   const balance = data?.balance ?? "0";
-  const { watch } = useFormContext<{
-    userInput?: string;
-    amount?: string;
-  }>();
 
   const isLedger = wallet?.type === "ledger";
   const onClick = () => {
     if (isLedger) return;
-    navigate(`/erc20/send/${asset.chainId}/${asset.address}`, {
-      state: {
-        step: "details",
-        form: {
-          userInput: watch("userInput"),
-          amount: watch("amount"),
-        },
-      },
-    });
-    toggleShow();
+    navigate(`/erc20/send/${asset.chainId}/${asset.address}`);
   };
 
   return (
@@ -49,22 +29,25 @@ export default function Erc20SelectItem({
       {balance !== "0" && (
         <button
           type="button"
-          className="flex items-center justify-between rounded-lg px-4 py-2 text-base font-medium text-daintree-200 hover:bg-daintree-700"
+          className="flex items-center justify-between rounded-lg px-3 py-2 text-base font-medium text-daintree-200 hover:bg-daintree-700"
           onClick={onClick}
         >
           <div
             className={twMerge(
-              "flex items-start gap-2",
+              "flex items-start gap-3",
               isLedger && "opacity-40",
             )}
           >
             <Layer2AssetImage
               tokenImage={asset.image ?? kasIcon}
-              tokenImageSize={24}
-              chainImageSize={20}
               chainImage={getChainImage(asset.chainId)}
             />
-            <span>{asset.symbol}</span>
+            <div className="flex flex-col items-start">
+              <span>{asset.symbol}</span>
+              <span className="text-xs text-daintree-400">
+                {textEllipsis(asset.address)}
+              </span>
+            </div>
           </div>
           {!isLedger && <span>{formatToken(parseFloat(balance))}</span>}
           {isLedger && (
