+++ components/send/asset-selector/EvmKasSelectItems.tsx
@@ -4,11 +4,7 @@ import { NetworkType } from "@/contexts/SettingsContext";
 import EvmKasSelectItem from "./EvmKasSelectItem";
 import { numberToHex } from "viem";
 
-export default function EvmKasSelectItems({
-  toggleShow,
-}: {
-  toggleShow: () => void;
-}) {
+export default function EvmKasSelectItems() {
   const [settings] = useSettings();
   const network = settings?.networkId ?? NetworkType.Mainnet;
   const supportedChains =
@@ -20,7 +16,6 @@ export default function EvmKasSelectItems({
         <EvmKasSelectItem
           key={numberToHex(chain.id)}
           chainId={numberToHex(chain.id)}
-          toggleShow={toggleShow}
         />
       ))}
     </>
