+++ components/send/evm/evm-kas-send/EvmKasSend.tsx
@@ -49,7 +49,7 @@ export default function EvmKasSend() {
   const onBack = () => {
     setStep((prevState) => {
       if (steps.indexOf(prevState) === 0) {
-        navigate("/dashboard");
+        navigate(-1);
       }
 
       const stepIdx = steps.indexOf(prevState);
