+++ components/send/kas-send/KasSend.tsx
@@ -59,7 +59,7 @@ export default function KasSend() {
   const onBack = () => {
     setStep((prevState) => {
       if (steps.indexOf(prevState) === 0) {
-        navigate("/dashboard");
+        navigate(-1);
       }
 
       const stepIdx = steps.indexOf(prevState);
