+++ entrypoints/popup/router.tsx
@@ -77,6 +77,7 @@ import Erc20Send from "@/components/send/evm/erc20-send/Erc20Send";
 import SelectAddress from "@/components/screens/receive-addresses/SelectAddress";
 import KaspaReceiveAddress from "@/components/screens/receive-addresses/KaspaReceiveAddress";
 import EvmReceiveAddress from "@/components/screens/receive-addresses/EvmReceiveAddress";
+import AssetSelect from "@/components/send/asset-selector/AssetSelect";
 
 const loadKaspaWasm = async () => {
   await init({ module_or_path: kaspaModule });
@@ -199,6 +200,7 @@ export const router = createHashRouter([
                 loader: keyringGuard,
                 children: [
                   // Sending
+                  { path: "asset-select", element: <AssetSelect /> },
                   { path: "token-transfer", element: <Krc20Transfer /> },
                   { path: "kas/send", element: <KasSend /> },
                   { path: "krc20/send/:tick", element: <Krc20Send /> },
